#266: reworking Move Counts Viewer code (still not working)
