+++ src/lbaf/Applications/conf.yaml
@@ -82,3 +82,4 @@ generate_meshes:
   y_ranks: 2
   z_ranks: 1
   object_jitter: 0.5
+generate_multimedia: True
